requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pypdf==4.1.0
faiss-cpu==1.8.0
numpy==1.26.4
sentence-transformers==2.6.1
requests==2.31.0
